move annotation meta utils
