+++ lib/annotation/meta.js
