+++ lib/request/browser-swarm.js
@@ -1,6 +1,6 @@
 const {createBrowserSwarm} = require('hyperswarm-ws')
 const {AbstractRequestSwarm} = require('./abstract-swarm')
-const {parseId} = require('../util/annotation')
+const {parseId} = require('../annotation/meta')
 
 const gatewayUrls = ['wss://hyperswarm-ws-gateway.kassel.works']
